refactor(-): :hammer: altera nome de classes internas
