+++ src/dieta_especial/solicitacao_dieta_especial/api/viewsets.py
@@ -100,12 +100,10 @@ from src.dieta_especial.tasks import (
     gera_xlsx_relatorio_historico_dietas_especiais_async,
     gera_xlsx_relatorio_recreio_nas_ferias_async,
 )
-from src.dieta_especial.utils import (
+from src.dieta_especial.utils_ import (
     filtra_relatorio_recreio_nas_ferias,
-    filtrar_alunos_com_dietas_nos_status_e_rastro_escola,
     gera_dicionario_historico_dietas,
     gerar_filtros_relatorio_historico,
-    trata_lotes_dict_duplicados,
 )
 from src.escola.models import (
     Aluno,
@@ -127,6 +125,10 @@ from src.relatorios.relatorios import (
 )
 
 from .pagination import RelatorioPagination
+from .utils import (
+    filtrar_alunos_com_dietas_nos_status_e_rastro_escola,
+    trata_lotes_dict_duplicados,
+)
 
 
 class SolicitacaoDietaEspecialViewSet(
