+++ src/dieta_especial/tasks/relatorio_recreio_nas_ferias_xlsx.py
@@ -13,7 +13,7 @@ from src.dieta_especial.tasks.utils.relatorio_recreio_nas_ferias import (
     gera_dicionario_relatorio_recreio,
     gera_xlsx_relatorio_recreio_nas_ferias,
 )
-from src.dieta_especial.utils import (
+from src.dieta_especial.utils_ import (
     filtra_relatorio_recreio_nas_ferias,
 )
